Deprecate Thenable and alias to PromiseLike
TypeScript has a builtin type for promise-like values: PromiseLike.
Thenable attempts to achieve the same result, but it has some
compatibility issues.

This change aliases Tnenable to PromiseLike, to make it compatible with
the rest of the TypeScript ecosystem.

Additionally this deprecates Thenable. There is no practical use case
for it, but it is confusing to communicate something other than
PromiseLike to users. All occurrences of Thenable were replaced with
PromiseLike.